Event number from 32 to 64 bits in Core
Change the event number in the EventID from 32 bits to
64 bits. This includes only changes for the Core part
of CMSSW. There are a couple hundred places outside the
Core code where this value is being assigned to types
with less than 64 bits that will also need changes to
work properly with data that has an event number exceeding
32 bits. Includes tests for the changes in the Core code.

+++ DataFormats/FWLite/test/test.cppunit.cpp
@@ -431,7 +431,7 @@ void testRefInROOT::testThinning() {
     // It is just checking that we read the values known to be
     // be put in by the relevant producer.
 
-    int offset = 100 + 100 * events.eventAuxiliary().event();
+    int offset = static_cast<int>(100 + 100 * events.eventAuxiliary().event());
 
     // In the D branch this tests accessing a value in
     // thinned collection made from a thinned collection
@@ -567,7 +567,7 @@ void testRefInROOT::testThinning() {
     // It is just checking that we read the values known to be
     // be put in by the relevant producer.
 
-    int offset = 100 + 100 * multiChainEvents.eventAuxiliary().event();
+    int offset = static_cast<int>(100 + 100 * multiChainEvents.eventAuxiliary().event());
 
     // In the D branch this tests accessing a value in
     // thinned collection made from a thinned collection
